✨ Add upload file

+++ backend/src/main/java/com/nx/lifesyncbackend/common/utils/SecurityUtils.java
@@ -44,7 +44,7 @@ public class SecurityUtils {
     public String getUserJwt(User user) {
         HashMap<String, Object> map = new HashMap<>();
         map.put("user", user);
-        map.put("expire_time", (Long) (System.currentTimeMillis() + expire));
+        map.put("expire_time", System.currentTimeMillis() + expire);
         return JWTUtil.createToken(map, salt.getBytes());
     }
